[AMD-SMI] Refactor py-interface: extract utilities and relocate handle helpers
- Create amdsmi_interface_utils.py with 14 utility functions + constants
  extracted from amdsmi_interface.py (format_bdf, parse_bdf, check_res, etc.)
- Move get_*_handles helper methods from py-interface to AMDSMIHelpers
  class in amdsmi_cli, using public amdsmi_get_processor_handles_by_type API
- Move amdsmi_get_processor_handles_by_type C declaration from
  tagEsmiProcDiscovery (#ifdef ENABLE_ESMI_LIB) to tagProcDiscovery
  (unconditional) in amdsmi.h
- Move amdsmi_get_processor_handles_by_type implementation outside
  #ifdef ENABLE_ESMI_LIB guard in amd_smi.cc
- Add amdsmi_interface_utils.py to py-interface CMakeLists.txt build
- Update all callers in amdsmi_commands.py and amdsmi_helpers.py
